Create a ReadMe file.
The purpose of this readme file is to give information about what the whole project was about and what the initial requirements are. I chose to make a blog site on the scope of Environment because it is an area that interests me and I want to continue on this project even after the submission of the project.